+++ interface/INetIO.h
@@ -19,12 +19,12 @@ namespace hudp {
         virtual bool Bind(uint64_t socket, const std::string& ip, uint16_t port) = 0;
 
         // send a msg to client
-        virtual int SendTo(uint64_t socket, const char * buf, int len, const std::string& ip, uint16_t port) = 0;
-        virtual int SendTo(uint64_t socket, const char * buf, int len, const std::string& ip_port) = 0;
-        virtual int SendTo(uint64_t socket, const char * buf, int len) = 0;
+        virtual int SendTo(uint64_t socket, const char * buf, uint32_t len, const std::string& ip, uint16_t port) = 0;
+        virtual int SendTo(uint64_t socket, const char * buf, uint32_t len, const std::string& ip_port) = 0;
+        virtual int SendTo(uint64_t socket, const char * buf, uint32_t len) = 0;
 
         // receiver msg from socket
-        virtual int RecvFrom(uint64_t sockfd, char *buf, size_t len, std::string& ip, uint16_t& port) = 0;
+        virtual int RecvFrom(uint64_t sockfd, char *buf, uint32_t len, std::string& ip, uint16_t& port) = 0;
 
         // return a udp socket if success. else return 0
         virtual uint64_t UdpSocket() = 0;
